Pushed noun resolving down the line
to the UriResolver felt more logical. Also pass the functionName to the
querystring resolver. After some debate thought it was suposed to get it
aswell. Fixed the flickr example AGAIN.... looks better now. Corrected a
typo. Did some clean up. Updated the version in the AssemblyInfo and
nuspec. Updated the readme.